requirements.txt
awsebcli==3.14.7
blessed==1.15.0
boto3==1.9.63
botocore==1.12.63
cached-property==1.5.1
cement==2.8.2
certifi==2018.10.15
chardet==3.0.4
colorama==0.3.9
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.3
django-filter==2.0.0
django-rest-swagger==2.2.0
djangorestframework==3.9.0
docker==3.6.0
docker-compose==1.21.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
idna==2.6
itypes==1.1.0
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
Markdown==3.0.1
MarkupSafe==1.1.0
openapi-codec==1.3.2
pathspec==0.5.5
python-dateutil==2.7.5
pytz==2018.7
PyYAML==3.13
requests==2.18.4
s3transfer==0.1.13
semantic-version==2.5.0
simplejson==3.16.0
six==1.11.0
spotipy==2.4.4
termcolor==1.1.0
texttable==0.9.1
uritemplate==3.0.0
urllib3==1.22
wcwidth==0.1.7
websocket-client==0.54.0
